make projects section responsive and separate mobile from desktop logic
