+++ src/components/projects/Projects.styles.tsx
@@ -1,20 +1,20 @@
 import { makeStyles, Theme } from "@material-ui/core/styles";
 
 const useStyles = makeStyles((theme: Theme) => ({
-  sectionDesktop: {
+  sectionProjects: {
+    paddingTop: "40px"
+  },
+  projectsDesktop: {
     display: "none",
     [theme.breakpoints.up("md")]: {
       display: "flex"
     }
   },
-  sectionMobile: {
+  projectsMobile: {
     display: "flex",
     [theme.breakpoints.up("md")]: {
       display: "none"
     }
-  },
-  sectionServices: {
-    paddingTop: "40px"
   }
 }));
